git-svn-id: http://www.ifountain.org/repos/os/RapidModules@2620 a08bcbf9-5f6c-4d33-af0a-e494bc2cacca

+++ RapidCMDB/grails-app/conf/BootStrap.groovy
@@ -46,7 +46,7 @@ class BootStrap {
     {
         RapidStringUtilities.registerStringUtils();
         Date.metaClass.'static'.now = {
-            return System.currentTimeMillis()/1000;
+            return (long)(System.currentTimeMillis()/1000);
         }
     }
